+++ lizard_history/models.py
@@ -57,6 +57,7 @@ def pre_save_handler(sender, instance, **kwargs):
 
 @receiver(models.signals.post_save)
 def post_save_handler(sender, instance, **kwargs):
+    print sender
     if _is_monitored(sender):
         kwargs.update(signal_name='post_save')
         handlers.db_handler(sender, instance, **kwargs)